\name{rcpps-package}
\alias{rcpps-package}
\alias{rcpps}
\docType{package}
\title{
\packageTitle{rcpps}
}
\description{
\packageDescription{rcpps}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{rcpps}
\packageIndices{rcpps}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{rcpps}

Maintainer: \packageMaintainer{rcpps}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file ~~
~~ KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
